Onix_web/requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.1
django-crispy-forms==1.9.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
flower==0.9.4
future==0.18.2
httpie==2.1.0
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.1
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
mysqlclient==1.4.5
packaging==20.4
Pillow==7.1.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
